requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
supabase==2.9.1
twilio==9.3.0
bcrypt==4.2.0
httpx==0.27.2
pydantic==2.9.0
